support for installing as a windows service
