+++ fliplet-agent.js
@@ -6,5 +6,7 @@ const program = require('commander');
 
 program
   .version(package.version)
-  .command('start [path/to/config.js] [--test]', 'Start the agent. Use the --test command to perform a dry run and avoid writing data to Fliplet servers.')
+  .command('start [path/to/config.js] [--test]', 'Start the agent in foregroud mode. Use the --test command to perform a dry run and avoid writing data to Fliplet servers.')
+  .command('install [path/to/config.js]', 'Install a config file to run as a background service (Windows only).')
+  .command('uninstall [path/to/config.js]', 'Uninstall a background service.')
   .parse(process.argv);
