+++ libs/logger.js
@@ -8,6 +8,16 @@ function getDate() {
   return moment().format('YYYY-MM-DD HH:mm:ss');
 }
 
+  // Windows Service Logger to Event Viewer
+if (process.env.SERVICE) {
+  const EventLogger = require('node-windows').EventLogger;
+  const winlog = new EventLogger('flipletagent.exe');
+
+  console.info = winlog.info;
+  console.error = winlog.error;
+  console.warn = winlog.warn;
+}
+
 logger.error = function (message) {
   console.error(chalk.red(`[${getDate()}]`), chalk.yellow('[ERROR]'), chalk.red(message));
 }
